Add title prop to Layout component
The code updates included adding a new 'title' prop to the Layout component in addition to the existing 'lang' prop. Now, the title can be customized per page based on the prop value. However, if no value is provided, the fallback title functionality based on language ('lang' prop) is still maintained.