Implement multilanguage support for navigation labels
The commit introduces multilanguage support for navigation labels by dynamically translating them based on URL language code. This was achieved by setting up language keys and a function to translate labels to either English or Spanish based on the user's selected language. This results in a more easily localized and user-friendly navigation interface.